[EST-101] Refactor : 상수, Enum 클래스는 common 패키지로 이동
