+++ src/main/java/com/blackshoe/esthete/common/Gender.java
@@ -1,4 +1,4 @@
-package com.blackshoe.esthete.entity;
+package com.blackshoe.esthete.common;
 
 public enum Gender {
     MALE, FEMALE
