+++ src/main/java/com/blackshoe/esthete/common/Role.java
@@ -1,4 +1,4 @@
-package com.blackshoe.esthete.entity;
+package com.blackshoe.esthete.common;
 
 public enum Role {
     USER("ROLE_USER");
